// frontend/src/shared/api/model/updateCircuitStopRequest.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Tarikma API
 * Backend API for Tarik.ma
 * OpenAPI spec version: v1
 */

export interface UpdateCircuitStopRequest {
  position?: number;
  /** @minimum 1 */
  dayNumber?: number;
  /**
   * @minLength 0
   * @maxLength 20
   */
  stopKind?: string;
  /**
   * @minLength 0
   * @maxLength 20
   */
  mealType?: string;
  startTime?: string;
  endTime?: string;
  /** @minimum 0 */
  durationMinutes?: number;
  notes?: string;
}